setup.py: Mention the 3-clause BSD license in our metadata
I don't know if PyPI has a convention for naming these, but that's
what the Open Source Initiative calls this one [1].  However, the text
is a bit different from the official upstream version which has:

  3. Neither the name of the copyright holder nor the names of its
     contributors may be used...

instead of the local:

  * The names of its contributors may not be used...

There's also an "COPYRIGHT HOLDER" -> "COPYRIGHT OWNER" change in the
local disclaimer.

It would be nice if we could stick to the standard upstream phrasing.

[1]: http://opensource.org/licenses/BSD-3-Clause