Arrange db fixtures in a more extensible way
The current scheme is not conducive to adding additional modifier
fixtures similar to ``django_db_reset_sequence``, such as a fixture for
``serialized_rollback`` or for specifying databases.

Instead, arrange it such that there is a base helper fixture
`_django_db_helper` which does all the work, and the other fixtures
merely exist to modify it.